Updates examples
