--- index.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node --harmony
-
-require('./lib/main')();
